+++ src/syntax/mod.rs
@@ -267,16 +267,26 @@ impl Lexer {
         gs: &mut GlobalState,
     ) -> Option<Rect> {
         let modal = self.modal.as_mut()?;
-        let maybe_pos = self.modal_rect.and_then(|rect| {
+        let found_positon = self.modal_rect.and_then(|rect| {
             let row = gs.editor_area.row + relative_editor_position.row;
             let column = gs.editor_area.col + relative_editor_position.col;
             rect.relative_position(row, column)
         });
-        // click out of modal or click finished modal
-        if !matches!(maybe_pos, Some(position) if !modal.mouse_click_and_finished(position, &self.lang, gs)) {
-            self.modal.take();
-        };
-        self.modal_rect.take()
+        match found_positon {
+            // click outside modal
+            None => {
+                self.modal.take();
+                self.modal_rect.take()
+            }
+            Some(position) => match modal.mouse_click_and_finished(position, &self.lang, gs) {
+                // modal finished
+                true => {
+                    self.modal.take();
+                    self.modal_rect.take()
+                }
+                false => self.modal_rect.take(),
+            },
+        }
     }
 
     pub fn mouse_moved_modal_if_exists(&mut self, row: u16, column: u16) -> Option<Rect> {